Melonking, I love you
melonking.et has opened my eyes to how powerful iframes actually are, and it is for this reason that I'm redesigning my site again. I'm doing it now because I don't want to sunk cost fallacy my way into making the full site and then instantly regretting it and making this one after that.