Query strings now change when link is clicked in the iframe
This was a relatively clean solution. The biggest issue is that all my iframe pages need to link to a script, but chances are, I'm gonna do something in the future that would need this anyways. If that's the case, then it's better to add that requirement when there are 3 pages and not 300.