RAMBLINGS: Restarting <wavy> implementation.
I thought this whole time that Marked was trying to get rid of my custom tags by using character references in place of arrows.
But, to the surprise of effing nobody, I am a moron. Markdown was only modifying the **invalid** custom tags, and the well-formed custom tags? They work perfectly! Nothing is wrong!

I have spent the last month going insane over what was, in the end, a mere misunderstanding. I know it is bad to self-deprecate, but how can I **NOT** call myself a moron?????